backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.31
alembic==1.13.2 
PyMuPDF==1.24.2 
python-multipart==0.0.9  
chromadb==0.4.24

openai==1.35.3  
google-generativeai==0.5.4  
tiktoken==0.7.0  

serpapi==0.1.5
brave-search==0.1.7  # Example package for Brave API (if official client is not available, fallback to requests)

pydantic==2.8.2
pydantic-settings==2.2.1
requests==2.32.3
numpy==1.26.4
pandas==2.2.2

fastapi-cors==0.0.6  
loguru==0.7.2
